feat(bazar-list-dynamic): get data asynchronously
